requirements.txt
langchain
google-generativeai
PyPDF2
streamlit
chromadb
python-dotenv